Preallocate decompress destination buffer
This patch updates Encode to store the uncompressed length as a varint
at the start of the buffer.  This allows us to preallocate the
destination when decoding and move all the append() calls to plain
assignments or calls to copy().  This is a serious optimization win,
making decoding approximately 20% faster.

This change to the output format means we are breaking interop with
other implementations.  However, many implementations (java/python)
already do exactly this, so in some sense we're _increasing_ our
interoperability, we just happen to be making an incompatible change.