Detect more out-of-bounds errors
Previously we would panic due to an array-of-bounds error.  It's nicer
to return ErrCorrupt instead.

Note that because Go is a memory-safe language, we're not going to hit
any of the problems the recently publicized integer-overflow error might
introduce, we would just panic instead.  Except that now we don't panic,
we return an error.